pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "DataRPG"
version = "0.1.0"
description = "A data-driven attribute management system for RPGs"
readme = "README.md"
authors = [{ name = "Felipe Cardoso", email = "seu-email@exemplo.com" }]
requires-python = ">=3.8"
license = { text = "MIT" }
dependencies = []
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Games/Entertainment :: Role-Playing",
]

[project.optional-dependencies]
yaml = ["PyYAML>=6.0"]
toml = ["toml>=0.10.2"]
all = ["PyYAML>=6.0", "toml>=0.10.2"]

[project.urls]
"Homepage" = "https://github.com/felCardoso/dataRPG"
"Bug Tracker" = "https://github.com/felCardoso/dataRPG/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
